+++ extensions/kagi-news/CHANGELOG.md
@@ -3,9 +3,9 @@
 ## [1.3] - {PR_MERGE_DATE}
 
 - Migrated to the new API, which Kagi News is now focusing on, so no more use of the static JSON files
-- New 'Time Travel' command to access Kagi News archives for a specific date
+- Added a new 'Time Travel' command to access Kagi News archives for a specific date, thanks to the new API
 - New favorite system for the 'Daily News' command to pin your categories of interest at the top of the Cmd+P menu
-- Improve the main article view so that the listed references link directly to the articles, similar to the listed sources in the right column
+- Revamped the listed sources with minimalist, clickable numbered superscript references that link directly to each source
 - Added the missing Simplified Chinese language 
 - Renamed the "On this Day" section to "Today in History", as it appears on the official website